requirements.txt
pandas
email-validator
tkinterweb
python-dotenv
ruff